Bump `markdown-to-jsx` >= `7.3.0`
Bumping to version `7.3.0` of `markdown-to-jsx`, which fixes a bug regarding import statements being used outside of modules. This was originally reported under #9 (but closed due to being a `markdown-to-jsx` issue and not a `mui-markdown` issue).

See the released version (https://github.com/probablyup/markdown-to-jsx/releases/tag/v7.3.0) which includes the patch for the bug (https://github.com/probablyup/markdown-to-jsx/pull/500).